+++ bin/faucet/src/main.rs
@@ -47,7 +47,8 @@ const COMPONENT: &str = "miden-faucet-server";
 
 const ENV_API_BIND_PORT: &str = "MIDEN_FAUCET_API_BIND_PORT";
 const ENV_API_PUBLIC_URL: &str = "MIDEN_FAUCET_API_PUBLIC_URL";
-const ENV_FRONTEND_URL: &str = "MIDEN_FAUCET_FRONTEND_URL";
+const ENV_FRONTEND_BIND_PORT: &str = "MIDEN_FAUCET_FRONTEND_BIND_PORT";
+const ENV_NO_FRONTEND: &str = "MIDEN_FAUCET_NO_FRONTEND";
 const ENV_NETWORK: &str = "MIDEN_FAUCET_NETWORK";
 const ENV_NODE_URL: &str = "MIDEN_FAUCET_NODE_URL";
 const ENV_TIMEOUT: &str = "MIDEN_FAUCET_TIMEOUT";
@@ -126,17 +127,21 @@ pub enum Command {
         #[clap(flatten)]
         config: ClientConfig,
 
-        /// Port to bind the API server. The URL will be constructed as `http://0.0.0.0:<api-port>`.
-        #[arg(long = "api-bind-port", value_name = "PORT", env = ENV_API_BIND_PORT)]
+        /// Port to bind the API server.
+        #[arg(long = "api-bind-port", value_name = "PORT", env = ENV_API_BIND_PORT, default_value = "8000")]
         api_bind_port: u16,
 
-        /// Public URL to access the API server. If not set, will default to http://localhost:<api-bind-port>.
-        #[arg(long = "api-public-url", value_name = "URL", env = ENV_API_PUBLIC_URL)]
-        api_public_url: Option<Url>,
+        /// Public URL to access the API server.
+        #[arg(long = "api-public-url", value_name = "URL", env = ENV_API_PUBLIC_URL, default_value = "http://localhost:8000")]
+        api_public_url: Url,
 
-        /// URL to bind the frontend server. If not set, the frontend will not be served.
-        #[arg(long = "frontend-url", value_name = "URL", env = ENV_FRONTEND_URL)]
-        frontend_url: Option<Url>,
+        /// Port to bind the frontend server.
+        #[arg(long = "frontend-bind-port", value_name = "PORT", env = ENV_FRONTEND_BIND_PORT, default_value = "8080")]
+        frontend_bind_port: u16,
+
+        /// Optionally disable the frontend server.
+        #[arg(long = "no-frontend", value_name = "BOOL", default_value_t = false, env = ENV_NO_FRONTEND)]
+        no_frontend: bool,
 
         /// The maximum amount of assets' base units that can be dispersed on each request.
         #[arg(long = "max-claimable-amount", value_name = "U64", env = ENV_MAX_CLAIMABLE_AMOUNT, default_value = "1000000000")]
@@ -326,7 +331,8 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
                 },
             api_bind_port,
             api_public_url,
-            frontend_url,
+            no_frontend,
+            frontend_bind_port,
             max_claimable_amount,
             pow_secret,
             pow_challenge_lifetime,
@@ -407,14 +413,10 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
             let api_id = tasks.spawn(api_server.serve(api_url.clone())).id();
             tasks_ids.insert(api_id, "api");
 
-            if let Some(frontend_url) = frontend_url {
-                let api_local_url = Url::parse(&format!("http://localhost:{api_bind_port}"))?;
+            if !no_frontend {
+                let frontend_url = Url::parse(&format!("http://0.0.0.0:{frontend_bind_port}"))?;
                 let frontend_id = tasks
-                    .spawn(serve_frontend(
-                        frontend_url,
-                        api_public_url.unwrap_or(api_local_url),
-                        node_endpoint.to_string(),
-                    ))
+                    .spawn(serve_frontend(frontend_url, api_public_url, node_endpoint.to_string()))
                     .id();
                 tasks_ids.insert(frontend_id, "frontend");
             }
@@ -592,10 +594,10 @@ mod tests {
         let result = Box::pin(run_faucet_command(Cli::parse_from([
             "miden-faucet",
             "start",
-            "--api-bind-url",
-            "http://localhost:8000",
-            "--frontend-url",
-            "http://localhost:8081",
+            "--api-bind-port",
+            "8000",
+            "--frontend-bind-port",
+            "8081",
             "--node-url",
             stub_node_url.to_string().as_str(),
             "--store",
@@ -734,8 +736,10 @@ mod tests {
                     command: crate::Command::Start {
                         config,
                         api_bind_port,
-                        api_public_url: None,
-                        frontend_url: Some(Url::parse(frontend_url).unwrap()),
+                        api_public_url: Url::parse(&format!("http://localhost:{api_bind_port}"))
+                            .unwrap(),
+                        frontend_bind_port: 8080,
+                        no_frontend: false,
                         max_claimable_amount: 1_000_000_000,
                         api_keys: vec![],
                         pow_secret: "test".to_string(),